fix(ASan): superinstruction predecode heap-buffer-overflow + 2 link fixes
Found by running the test suite under ASan+UBSan:

1. superinstruction.hpp predecode() heap-buffer-overflow (CRITICAL):
   - new_code buffer was allocated as bc->length bytes, but fusion
     GROWS the code (LOAD_CONST_INT(3) + IADD(1) = 4 bytes →
     LOAD_CONST_INT__IADD(5) = 5 bytes, a 1-byte growth).
   - The comment 'fusion never grows code' was wrong.
   - Fix: allocate bc->length + 64 bytes.
   - Also fixed the destination bounds check that was incorrectly
     truncating the verbatim copy when new_pc exceeded bc->length
     (copy_len was set to 0, causing RETURN_VALUE to be dropped).
   - Also added bounds check in the branch-rewrite loop before
     read_op2/write_op2 to prevent OOB read/write.

2. test_jit_deopt SEGV in vtx_dispatch_jit (pre-existing, not fixed):
   - T1 JIT crashes on heap-object field access (LOAD_FIELD).
   - Reproducer: tests/opcode/test_jit_deopt.c test_jit_field_access
   - The JIT-compiled 'distance(Point)' method segfaults when
     accessing p.x + p.y. This is a real T1 codegen bug in
     LOAD_FIELD handling — needs separate investigation.

Link fixes (build-asan was failing to link several targets):
- Added vortex_trace, vortex_region, vortex_midtier, vortex_sota
  to test targets that were missing them (orchestrator.c calls
  vtx_trace_retrace_* which lives in vortex_trace).
- Fixed bench_arith, bench_jit_exec, bench_real_workload link lists.
- Bulk-fixed all test/benchmark targets missing vortex_trace.

Also fixed:
- tools/disassembler_main.c: uint16_t conversion warnings
- tests/unit/test_assembler_prog.c: uint16_t conversion warnings
- tests/regression/osr/test_osr18_atomic_dedup.c: sign-conversion
- tests/regression/osr/test_osr26_gc_conservative_scanner.c: sign-conversion
- src/main_new.c: uint16_t conversion in v2 header parse

Build: clean with -Werror under ASan+UBSan.
Tests: 79/81 pass under ASan+UBSan (2 remaining: test_pgo_sprint1
pre-existing, test_jit_deopt = real JIT field-access bug to fix next).

+++ src/main_new.c
@@ -198,8 +198,8 @@ static vtx_bytecode_t *load_bytecode_file(const char *filename, vtx_arena_t *are
             fclose(f);
             return NULL;
         }
-        hdr_max_locals = ((uint16_t)buf[0] << 8) | buf[1];
-        hdr_max_stack  = ((uint16_t)buf[2] << 8) | buf[3];
+        hdr_max_locals = (uint16_t)((((uint16_t)buf[0] << 8) | (uint16_t)buf[1]));
+        hdr_max_stack  = (uint16_t)((((uint16_t)buf[2] << 8) | (uint16_t)buf[3]));
     }
 
     /* Allocate bytecode structure */
